Renaming "View"
Rather than having `View` be a type, make `Dials` the actual root type
and have `View` be a method that returns the current "view" of the
configuration.

This will be followed up with another PR that introduces `Fill` which
will take a pointer to a config structure to eliminate having to do the
type assert.