flaghelper: add TimeWrapper flag.Value implementation for time.Time
This is needed primarily to avoid printing the zero time as the
default in PrintDefaults. Using a MarshalWrapper does not work
because its zero value wraps a nil encoding.TextUnmarshaler, which
returns an empty string from its String method, which does not match
the value of String for the zero time.Time.