Config: deepcopy the template value
Make a private copy of the default config (template) provided as an
argument to `Config` to avoid surprises when users modify the template
struct.

This is most important when a watching source is involved, as we now
have a long-lived reference to our template struct.

A natural way users will try to use the `Fill()` method (that we had
failed to anticipate) was to pass the template variable to that struct
directly.

Without this deep-copy, using `Fill()` in this way is:
 - thread unsafe, since the watcher assumes that value is immutible and
   hence is read without synchronization
 - surprising, since the current config becomes the default config,
   making removal of fields from config files default to the existing
   value, rather htan the normal default that was passed (as the
   base-layer) to dials.

Add a test verifying this behavior.