ez: cancel an inner context if watching disabled
If file-watching is not enabled, we should shutdown the monitor
goroutine when exiting the `ConfigFileEnvFlag` function.

Usually `dials.Config` is smart enough not to start a monitor when
there are no `Watcher` implementations in the source-list, but the
`Blank` source uses `Watcher` for its core functionality, so we need
to cancel the context passed to `Config` to actually clean up
resources.