OnNewConfig callback
Add a callback that's called just after installing a new configuration.

This eliminates the need for most use-cases to use an additional
goroutine to watch for config changes/updates.

The callback gets both the old and the new configs so clients can
differentially update other datastructures.

Since this new callback is executed by the callback manager goroutine,
and we try to keep callback execution serial, NewConfigHandler
implementations should be short to moderately long-running. (if they
take longer than the interval between updates to the config events
will back up)