dials: rename NewValue -> ReportNewValue
Per review comments, rename the NewValue on `WatchArgs` to
`ReportNewValue`.
@sergiosalvatore correctly pointed out that `NewValue` can be
mis-understood as a helper that will construct a new value for the
`Watcher`, rather than a way to report a new value.

The new name `ReportNewValue` matches the tentative names for other
methods that will be added to that interface in the next PR.