Improve Verification in EZ
It's common that we would have some required values that might only be
set in the config file.  Previously when using the EZ package, the
`VerifiedConfig` interface was running on the initial stacking before
the config file had a chance to load values in to the overall values.
This allows us to skip the initial verification and wait until the file
is (optionally) considered before running verification.