dials: add ViewVersion
The next change will introduce a way to register (and unregister)
callbacks at runtime, with active watchers. To decided whether or not to
immediately call the callback with the current config version. In
particular, it's likely that the caller used a specific version for
setting up the structs/objects that will be manipulated by the callback.
That config-version is the relevant version.

The new CfgSerial type includes a backpointer to the version it refers
to so if the new-config callback needs to deliver a new configuration,
because the "current" version may be later. This way we don't need to
hold onto all versions we've ever seen in order to deliver a catch-up
new-config callback.