dials: runtime callback (un)registration
In many processes that need callback streams, the component that will
consume the callbacks needs some initial configuration during
construction.
This produces a paradox when constructing the New Config callback to
insert into the Params struct.

To resolve this, one have two options:
 - Create a dummy callback with some synchronization and safety-checks
   so one can register the new callback with that other type later.
 - Update Dials so it supports runtime callback registration.

Notably, this implementation uses the new CfgSerial type to provide the
necessary safety-checks and guarantee that if one consumes config
version 3, but while object setup is ongoing, versions 4, 5, and 6 come
along, the callback will only be called with oldCfg from version 3, and
newCfg from version 6.

This guarantees that the callback is called on the latest version at
registration-time.

It's worth noting that since registration and unregistration are done on
the same channel as the event-delivery to the callback manager's
goroutine, we get explicit ordering and before/after relationships, so
we don't need to block on registration.

However, when unregistering, we must block until the unregister is
processed because the unregister event may be queued behind several
new-config events on the channel, which would otherwise be delivered
after the callback is unregistered.

There is one (preexisting) caveat: if we fill up the callback-event
channel (currently statically capacity 64), we drop newer callbacks in
favor of older ones. A future change should introduce some look-aside
saturation handling there.