Transformer: Construct the right recursive types
When recursively unmangling structs with ReverseTransform construct the
mangled type for interacting with the intermediate types.

This is a partial revert of 9ffa9007c3918ec9cb214094f6bd9dec7e2bdf2b.
I misunderstood the ordering of the code unmangling.

This commit adds a test that creates an incompatible struct for the
inner type of a slice (converting from 1 to 2 fields). That test failed
in an interesting way, and unmasked my mistake.

In maybeRecursivelyUnmangle the type has to be the mangled type so it
can be passed to Unmangle after it's recursed into the fields within
that field.

Note: this doesn't trigger in many existing test-cases because the
mangled type ends up being assignable to the unmangled type.