transform: propagate nil slice in SetSliceMangler
Also address a De Morgan's law nit from staticcheck in the guard

For dials' field-stacking to work, we need nil-slices to convert to
nil-maps. Add an explicit check so a nil/missing slice/list doesn't turn
into an empty map, which is semantically different in that Dials'
stacking code sees empty as "specified, but empty", while nil is
interpreted as "unspecified".

As a result, in the former case that field in the config is cleared, and
in the latter case it's inheritted from the next layer down. (possibly
the default in the template struct)