blank: add SetSourceNoVerify
There are times where one is setting multiple sources, and the
intermediate stages (with only a subset of sources installed) fails the
config's Verify() method. Make it possible for callers to explicitly
skip verification when setting a new blank source in the cases where
it's needed.

This is not expected to be a common case, but there are situations where
there are multiple configuration files that need to all be processed
(possibly in a specific order) before the configuration will become
valid.

It is unlikely that any watching sources besides the blank source will
need to use the new BlockingReportNewValueSkipValue, but it is now part
of the public API so Blank is not _that_ special.