Create SingleTypeSubstitutionMangler
Implement a Mangler implementation that takes two type-arguments that
are directly convertible to convert to and from.

This will facilitate point-substituting time.Duration with a type that's
also an int64, but has an `UnmarshalJSON` method that calls
`time.ParseDuration` if it's passed a JSON string, and parses an integer
as nanoseconds if passed a JSON integer/number. -- for use by the json
and cue decoders.