AliasMangler
Add a new mangler that allows us to have aliases for dials tags to
facilitate migrating from one name to another seamlessly. If both the
old name and the new name are set, an error will be returned because the
expectation is that you're using one or the other exclusively.