transform: teach flatten mangler to propagate tags
Add a way for the flatten mangler to be configured to propagate
particular tags as it descends.

Notably, this will be useful for the flag, pflag and env sources
so flags/env-vars can be disabled at a higher level when interacting
with types you don't control. (and in some cases allow one to override a
flag-name when embedding a type which will only generate a single flag)